f/161849 merge sort fixes and tests
