+++ packages/search/src/util/index.ts
@@ -1,0 +1,1 @@
+export { kMerge } from "./merge-sort/merge";
